requirements.txt
asgiref==3.4.1
Babel==2.9.1
dj-database-url==0.5.0
Django==3.1.13
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==3.0.2
django-oscar-api==3.0.0
django-phonenumber-field==3.0.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
factory-boy==2.12.0
Faker==8.9.1
gunicorn==20.1.0
phonenumbers==8.12.26
Pillow==8.3.0
psycopg2==2.7.7
purl==1.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
